Identified and addressed an issue with the toggle-favorite route:
We found that the id property was undefined when trying to remove a favorite.
We added the missing columns to the favorites model in db.js to match the database schema.
We updated the code to directly delete the record using user_id and card_id.
Implemented a feature to highlight favorite cards on the /all route:

We updated the /all route handler to fetch the user's favorite card IDs.
We marked the cards that are in the user's favorites.
We passed this information to the template and updated the star icon accordingly.
Overall, we've improved the functionality and user experience of your application by fixing a bug and adding a new feature.